Update revisionDesc
 - Submit - Entered text and metadata for Pap. Congr. XXVII p. 967-969 no. 1.
Changed the two occasions of (δραχμ.) on line five to (δραχμὰς) and (δραχμαὶ), in that order, since keeping the edition's format was interfering with the Leiden+, and those are the correct expansions for each occasion.
I think?
The second occasion is definitely nominative, and the second should be accusative. Both are in plural, though. (Ηλιάνα Γιαννούχου)
 - Vote - HGVAccept - Fine (James Cowey)
 - Finalized - Fine (James Cowey)