Update revisionDesc
 - Submit - I transcribed the text according to the edition of Nathan Carlig (Lina Elhage-Mensching)
 - General -  (Rodney Ast)
 - General - Add a blank space after the numbers in lines 1. and 3. (Rodney Ast)
 - Vote - Accept-Straight-to-Finalization - Text:  Add a blank space after the numbers in lines 1. and 3. Lower case letters at beginning of sentences in lines 1 and 3. Place dot below mu at very end of line 2; add dots to τ̣[α]ῦ̣τ̣α̣, l. 3; underdot needed under sigma in ἄγγελος in last line. Add lenis over ἰδοὺ
Metadata: Biblio 56829 would be the previous edition, 82132 is the principal. Add catalogue 12796, no. 325. Change display name to something that reveals the content of the text (the legacy display name can serve as the basis for this). TLG id for the New Testament is 0031. Lines = 4; use periods not commas for dimensions. For place, collection, etc., say "Location unknown; formerly Baarn, Private collection Moen 631" (Rodney Ast)
 - Finalized - Requested changes made. Ready (Rodney Ast)