Lines 2–3 of the verso were forgotten. In entered the text as it is in the edition, though the text is hardly correct (I shall comment on this elsewhere).
I am not sure, though, what to make of προκ at the beginning of l. 3. In Maspero's text, there is no abbreviation stroke and neither is there one on the photo. It appears that Maspero thought (the hand is not well trained) that the scribe wrote προκ of πρόκ(ειται) before remarking that he misses the essential formula. But then why is προκ not cancelled? At this position ὁ προκ(είμενος) should be expected, but again, there is no abbreviation stroke and this alternative is apparently not what Maspero had in mind.

 - Added ll. 2–3 from the edition (there were just forgotten)

Signed-off-by: Nikolaos Gonis <nikolaosgonis@googlemail.com>