Update revisionDesc
 - Submit - Lines 2–3 of the verso were forgotten. In entered the text as it is in the edition, though the text is hardly correct (I shall comment on this elsewhere).

I am not sure, though, what to make of προκ at the beginning of l. 3. In Maspero's text, there is no abbreviation stroke and neither is there one on the photo. It appears that Maspero thought (the hand is not well trained) that the scribe wrote προκ of πρόκ(ειται) before remarking that he misses the essential formula. But then why is προκ not cancelled? At this position ὁ προκ(είμενος) should be expected, but again, there is no abbreviation stroke and this alternative is apparently not what Maspero had in mind. (Matthias Stern)
 - General -  (Matthias Stern)
 - General - Actually, the edition has the lacuna in the last line run continuously στι[χῖ(*) μοι ὡς πρόκ](ειται), but I always got an Leiden+ Error, when I tried to do it that way (and I did not find this particular case (a lacuna extending over multiple words but starting and ending within words) in the help file. (Matthias Stern)
 - Vote - DEFER - Too complex for a PN entry? (Thomas Backhuys)
 - Vote - Accept-Straight-to-Finalization - Voting this through to get it processed. (Nikolaos Gonis)
 - Finalized - Ready. (Nikolaos Gonis)