Update revisionDesc
 - Commit - corrected line numbering from 82-end: line 82 (Λέωνι̣ Θεόδ̣οτος) was misnumbered as 81 (Elisa Nury)
 - Submit - Corrected line numbering. (Elisa Nury)
 - Vote - AcceptText - The new numbering is still not right, because, counting in the expunged text, the new line 82 would officially be line 87, and the new line 83, line 91.
My advice would be to keep to the decision of the editors to not count the expunged text in, and stick to the line numbers of the editors, and thus 
to correct the faulty text in PN:
Traces 5 lines
81Λέωνι̣ Θεόδ̣οτος
Traces ca.4 lines
into:
Traces of 10 expunged lines including (l. 6) ⟦Λέωνι̣ Θεόδ̣οτος⟧
and leave the rest as is. (Cisca Hoogendijk)
 - Vote - AcceptText - I agree with Cisca. To start renumbering and diverging from the printed text would cause citation problems.  (James Cowey)
 - Finalized - Ready. The online image was consulted before making the slight changes to lines between lin81 in col. vii and lin82 in col. viii. (James Cowey)