Update revisionDesc
 - Submit - Tiping mistake correction at l.19.
The text in papyri.info read (ἔτους) γ. whereas the edition princeps gives (ἔτους) λ. (l. λα̣ or λδ̣). The reading γ was not recorded in the BL, not to find in the BP, and not attributed in papyri.info. Numerically, γ cannot be the first letter of a pair. A year 31 or 34 of Ptolemy II is wholly compatible with the range of the Zenon archive, opposite to a year 3. On the papyrus, the letter resembles more to a λ. It was probably just a tiping mistake. (Lavinia Ferretti)
 - Vote - Accept-Straight-to-Finalization - Fine. A common number conversion problem. (Nikolaos Gonis)
 - Finalized - Ready. (James Cowey)