Update revisionDesc
 - Submit - Submitted new entry copied from PDF.  (Clinton Kinkade)
 - Vote - Accept-Straight-to-Finalization - l. 4: change φάκου to φακοῦ and φάχου to φαχοῦ
l. 5: change [·] at the end to [.]
l. 6: change φάκου to φακοῦ and φάχου to φαχοῦ
l. 9 and 10: I do not quite see the point in the overstroke over δραχμῶν. The drachma-sign has the same overstroke in O.Krok. II 221, 235, 287 and 312, but there it has not been reproduced in our transcription. 
l. 9: the ed. encloses the underdot within the square brackets in round brackets, to denote a letter may or may not have stood there. I do not know if and how this can be reproduced in the transcription here. 
l. 16: reverse the order of ἔρρωσο (regularized form) and ἔρρωσω (actual reading).  (Demokritos Kaltsas)
 - Finalized - Ready. (James Cowey)