Update revisionDesc
 - Commit - I HAVE ADDED THE EDITION OF THE SECOND DOCUMENT IN A TOMOS SYNKOLLESIMOS. NEW EDITION (nikos litinas)
 - Submit - I HAVE ADDED THE EDITION OF THE SECOND DOCUMENT IN A TOMOS SYNKOLLESIMOS. NEW EDITION

P.Corn. inv. II 33 is a petition from A.D. 153-154 addressed to Nikippos, nomarch of Antinoou polis from four alabaster workers (Didymos son of Hermaios, Tothes son of Onnophris, Papentos son of Kilikas, and Psenosiris son of Sarapion) for the ornamentation of the Caesareum constructed in Antinoopolis by Antonios Dios. They asked for their salary from the month Phamenoth. The document was published by N. Lewis in RecPap 3 (1964), 25-27 (= SB VIII 9904). The document formed part of a tomos synkollesimos. As Lewis notes, still attached at the right is the left edge, containing 2-3 letters per line, of the next document in the roll, written in a different hand. The sheet-join starts from 1.7 cm from the right edge of the papyrus and is 1 cm in breadth. The right sheet was glued below the left one. This document is also a petition addressed to the same nomarch, Nikippos and it is of a different case than the preceding one. Nikippos, the nomarch of Antinoou, is attested also in P.Oxy. XL 2941 and 2942.
Image in http://quod.lib.umich.edu/a/apis/x-3252/c2_33r.tif.
The remaining text reads as I have added here.

Commentary: In l. 2 after παρά we expect the name of the petitioner
 (nikos litinas)
 - Vote - Return-Straight-to-Sender - Here is this text again. (James Cowey)
 - Submit - Added transcription of a fragment, described, but not in ed. princ. and in SB. (nikos litinas)
 - Vote - Accept-Straight-to-Finalization - Checked on photo. Fr. a ll. 1-2 were read also in RecPap 3 (1964) 27. (Andrea Bernini)
 - Finalized - Ready. (Nikolaos Gonis)