entry of PSI XVII 1688
Notes: 
l. 1a. I was unable to indicate that this line, as the ed. pr. claimed, is written in the upper margin of the papyrus, on the right; consequently I used Leiden+ "vac.?lin" to marked the space between ll. 1a and 1 in the ed. pr., but I am not sure that this tag is the most appropriate.

 - Created from SoSOL template

Signed-off-by: James Cowey <james.cowey@urz.uni-heidelberg.de>