Update revisionDesc
 - Submit - Entered from PSI 17.
At l. 10 the editor writes ἱϲτορεῖ ἐν [τῷ x τοῦ Τρωϊκοῦ

I have a doubt for what concerns the missing numeral. I tried to use the tag <#[.?]=#> for the missing numeral, but it does not display properly in the preview; instead I used the tag <#x=#>, but I do not know if it is correct.
 (Marzia D'Angelo)
 - Vote - Accept-Straight-to-Finalization - META: culture is ‘literature’; TEXT: use msup for line 1a; for l. 10 use <#[.1-2]=#>; add space in l. 11;  (C. Michael Sampson)
 - Finalized - This is good to go. (C. Michael Sampson)