Update revisionDesc
 - Submit - In the context of a webinar on subliterary papyri taught in Summer Semester 2021 by me and H. Essler, I've entered this mathematical exercise. The text includes a new reading reported to me by J. Lougovaya: in IIb, l. 8, the correct reading for the fraction is certainly η ' (bis). The number is not executed very well and somewhat resembles a poorly written ζ, the interpretation of K.A. Worp. One-seventh, however, is not sought in any of the other operations and the (correct) amount recorded in the line, χκη, is that of one-eighth. (Rodney Ast)
 - Vote - Accept-Straight-to-Finalization - I suggest either changing the 'Mini' in the title to 'Miniature' or taking the word out. In general notes, perhaps instead of "Two are inscribed on three sides in five columns . . .", give more detailed descriptions? For example: I.a: uninscribed, I.b, II.a, II.b: fractions; III.a, III.b: uninscribed. 
Introduction: insert 'is' in 'There a number'.
Text. I'd number lines in each col. separately (i.e. starting each with 1).
I.b.i.5: this is the only instance of 1/4 transcribed as  d´, so change here or all other 1/4 fractions to  d´.
II.a.i.10: insert space between last two fractions: <#ι̅=1/10#> <#λ̅=1/30#>
II.b.8: should there be a space after the semicolon in ';from photo'?
Commentary to II.a.ii.27: instead of 'However, it cannot be true of both', simply explain that this is the quotient for 1/16 of 3400, whereas that for 1/15 should have been σκϛ 𐅷.
Commentary to II.b.10: in 'the sums in lines 9-15 are wrongly aligned' change 'sums' to 'quotients'.
Commentary to II.b.14: add a fraction tick to 1/4 in σνα δ. 
Otherwise, looks good. 
 (Julia Lougovaya)
 - Finalized - Requested changes made. Ready (Rodney Ast)