Update revisionDesc
 - Commit - l.4. I clearly see traces of a ν. Therefore I think I should enter as a proposal for correction as:  <:[.?]ντα=PN A. de Frutos García (from photo)|ed|[.?]τα:> (Alba de Frutos García)
 - Commit - l.5: there are visible traces of three letters belonging to a fifth line that were left untranscribed in the ed. pr. Ι entered enter an extra line as 5. [.?].3[.?] (Alba de Frutos García)
 - Commit - I have also proposed a ed. correction to l. 3 (Alba de Frutos García)
 - Submit - In l.3., the editor printed  ]ναπλευσντεει[  and included in the commentary "piuttosto ἀ]ναπλεύσ<α>ντε<ς> εἰ[ς che ἀ]νάπλευσ<ό>ν τε εἰ]ς."
Well, I read   ]ναπλευσητεεισ̣[, i.e, η instead of a ν ( just compare to the η in l.1). There are also visible traces of the σ. 
I propose to read this as ἀ]ναπλεύση<ι> τε εἰς̣[,  on two grounds: 1. η instead of ηι  was not uncommon (see www.trismegistos.org/texirr/209) and, with regards  to η for 3rd pers aor subj in the Zenon archive see, e.g., PSI 4 405, 5: ἀποβη (l.l. ἀποβῆ<ι>). 2.  ἀναπλέω with εἰς is well attested in the documentary record, see, e.g., P. Lond. 7 1945, 2, also a letter from Hierokles. I have then entered this a as a editorial correction.
In l. 4 l I clearly see traces of a ν. Entered <:[.?]ντα=PN A. de Frutos García (from photo)|ed|[.?]τα:> καὶ περι[.?] 
On the other hand, there are visible traces of three letters belonging to a fifth line that were left untranscribed in the ed. pr. I just enter an extra line as 5. [.?].3[.?] and numered the following lines accordingly. (Alba de Frutos García)
 - Vote - Accept-Straight-to-Finalization - The corrections to ll. 3 and 4 are fine, but they have already been made in very similar form by Clarysse in his file “New Editions; Addenda and Corrigenda to the Texts” (https://www.trismegistos.org/arch/add_corr_zenon.pdf). So they will have to be entered under his name. The missing line must be added, attributed to A. d. F. G.   (Demokritos Kaltsas)
 - Finalized - Fine. (Demokritos Kaltsas)