Update revisionDesc
 - Submit - Coptic Text
There is different text added between lines 5, 7 and 9. It is indicated in smaller font. (MªJesús Albarrán Martínez)
 - Vote - AcceptText - lin5bis=>ⲉⲓⲥ ϩⲏⲏⲧⲉ; lin10=> add a second space before  ⲓ(¨)ⲏⲣⲏⲙⲁⲓⲁⲥ; lin19=>ⲉⲓⲥ ϩⲏⲏⲧⲉ;
I am not satisfied with the interlineal text 2 encoded this way, because it breaks to meaning and the syntax of text 1, but I don't how to encode it otherwise... Next voter?  (Loreleï Vanderheyden)
 - Vote - AcceptText - Lin5=> ⲉⲛⲉⲥⲛⲏⲩ ; lin11=> <:ⲉⲧⲉⲡ[.2]ⲗⲟⲩⲗⲟⲩⲥ|alt|ⲉⲧⲉ ⲡ[ϣⲛ] ⲗⲟⲩⲗⲟⲩⲥ:> ; lin12=> ⲙ̄ⲡⲥⲟⲛ ; lin13=> ⲡⲣⲉⲥⲃⲩⲧⲉⲣⲟⲥ is considered as a proper name, not to be regularized ; lin16=> ⲧ<: (without space). I agree that the encoding of the interlineal text 2 is not completely clear, but to treat it as another text would not be satisfactory either. (Anne Boudhors)
 - Finalized - James and I have decided to keep the second text in the interlines using "bis" to distinguish the second text, in order to keep the original display of the ostracon.
Ready.
 (Loreleï Vanderheyden)