Update revisionDesc
 - Submit - The way the end of l. 4 is represented now, it implies that the correct result of the calculation is (on the papyrus) 2222 nom. 15 1/2 ker. but should be 2222 nom. 16 1/2 ker. This is, however, not the case: it is 2222 nom. 15 1/2 ker. on the papyrus, but it should amount to 2222 nom. 16 ker. 
(I changed it so that it now depicts the asterisk being after the fraction, so *in the preview* it looks correct now. However, in order to do this I had to write it in Leiden+ as if *only the fraction* (1/2) should be read as "16". Is this the only possibility to do this or is there a workaround that makes the entire "15 1/2" be read as "16"?) (Matthias Stern)
 - General - If |reg| is the appropriate markup (as opposed to |corr|, for outright scribal error), then it should be <:<#ιϛ=16#>|reg|<#ιε=15#> <#𐅵=1/2#>:>. I prefer |corr| tags, to be honest, in which case encode <:<#ιϛ=16#>|corr|<#ιε=15#> <#𐅵=1/2#>:>. (C. Michael Sampson)
 - Vote - AcceptText - See my previous comment. (C. Michael Sampson)
 - Vote - AcceptText - It's not standard practice to indicate wrong arithmetic in the apparatus, but why not? I've just doen the same with a text I submitted. (Nikolaos Gonis)
 - Finalized - This is good to go (C. Michael Sampson)