Update revisionDesc
 - Submit - editorial corrections in i 1-5; ii 1-5; I added line 5 in the first column accordingly
there is a rho after the first νο(μισμάτων) in line i 2 which is somehow invisible; also I am unsure whether I rendered the square brackets of that line correctly, maybe they should include the second νο(μισμάτων)?
In line i 1, νο(μίσματα) should maybe be resolved as νο(μισμάτων), as the new reading has διά where the ed. pr. ad an erased από 

 (Conrad Theilemann)
 - Vote - Accept-Straight-to-Finalization - Several things to fix at finalization. (Nikolaos Gonis)
 - Finalized - done (Nikolaos Gonis)