PSI XVII 1668 (first DCLP ed.)
Some notes: 
I tried to put ll. 259 etc. among brackets, but I didn’t succeed. Moreover, It could be useful to insert a blank space between the indication of lines and the text.
ll. 1-2. [ἐ]νυαλίῳ[ca.3] what I would like to indicate is [ἐ]νυαλίῳ[     ], like in the edd.pr., but I was not able to do it; the same is for the following line.
l. 3. The line would be indent, but if I try to do so, I have problems in the visualization of the text.



 - Created from SoSOL template

Signed-off-by: C. Michael Sampson <mike_sampson@umanitoba.ca>