Update revisionDesc
 - Submit - PSI XVII 1668 (first DCLP ed.)

Some notes: 
I tried to put ll. 259 etc. among brackets, but I didn’t succeed. Moreover, It could be useful to insert a blank space between the indication of lines and the text.
ll. 1-2. [ἐ]νυαλίῳ[ca.3] what I would like to indicate is [ἐ]νυαλίῳ[     ], like in the edd.pr., but I was not able to do it; the same is for the following line.
l. 3. The line would be indent, but if I try to do so, I have problems in the visualization of the text.

 (Roberta Carlesimo)
 - Vote - Accept-Straight-to-Finalization - Remove callouts from transcription, indent line 3. Review all metadata before finalization. (C. Michael Sampson)
 - Finalized - This is good to go. (C. Michael Sampson)