Update revisionDesc
 - Submit - PSI XVII 1667 (first DCLP ed.)
Some notes: 
I have problem in the digital organization of the text: the square brackets have to be aligned. 
Col. I 3-12, the black squares inserted in the text are used, according to the ed.pr., to indicate a stripe of ink dripped from the r. 2; they have to be aligned. 
col. I 1-2: the insertion in the interlinear space has to be indent
col. I 3-4: νεικήσεις has to be indent.
Col. I 12: [ -ca.?- ] ◼︎ ψυχὴ̣ν ποιήσαν(*)τες. According to the reconstruction of the ed.pr., we have to supposed that there are not lost characters in the gap: the line has to be indent, in order to align it to the interpretamentum in the previous line.
Col. I 18-19. According to the reconstruction of the ed.pr., we have not to suppose that in the gaps at the end of the lines were lost characters, even if the papyrus was broken at that point. I am not able to encode that situation.
col. I 20. The ed. pr. says that minimum 12 lines are lost, but there aren’t indication to encode that. 
col. II 7.   ̣ [ ̣ ] βωτια⟦ν⟧\ν/εί[ρῃ]. The first traces are in ekthesis. The position of \ν/ is the same of other corrections I indicated as ||interlinear: ||. I don’t know if it is better to use the same encoding method; \ν/ has to be aligned with ⟦ν⟧.
Col. II 8. ἐ̣δηλή\⟦η⟧/σα[ντο]. \⟦η⟧/ is aligned with ή; see what indicated regarding the previous line.
 (Roberta Carlesimo)
 - Vote - HGVAccept - Has to be pushed through for DCLP (James Cowey)
 - Finalized - Ready. (James Cowey)