PSI XVII 1673 (first ed. DCLP)
Some notes:
It is an astronomical table; so the columns have to be visualized side by side. 
In the number of lines, I indicates 41, considering the columns as part of a table, as in the ed. pr.
Is it possible to align the numbers indicated as in the ed. pr.?
In col. V is it possible to encode that the lines lost are ll. 7-21, as indicated in the ed.pr.

Under 𐆊 I need to insert a under dot, but I was not able to do it in: 
col. I 32, 40
col IV 22, 23
col. IV 32, 33, 34, 36, 38 (under the second 𐆊 of each line)

Col. III 40. I was not able to visualize [ϛ] μ̣[η λ]δ λ instead of [ϛ] μ̣[η] [λ]δ λ





 - Created from SoSOL template

Signed-off-by: C. Michael Sampson <mike_sampson@umanitoba.ca>