Update revisionDesc
 - Submit - PSI XVII 1670 (first DCLP ed.). 

Some notes: 
In the introduction, column width is indicated in ca. 30 characters. I use this datum to indicate the number of characters lost in the gaps. 
Col. II 1-3. The asteriskos are in the left margin a little bit dislocated under l. 1 and 2; in the ed. pr. they are indicated between ll. 1-2 and 2-3. I was not able to reproduce that. Moreover, I have problem in the visualization of the signs.
Col. II 2. σ vac. ? κιρτωσ[. In the ed. pr. is indicated that the vac. is probably due to a damaged fiber.
Col. II 3. I have problem to indicate that the integration is at the end of  l. 2.
Col. II 12. l. τιμησι ̣. I am not able to visualize a question mark after the word; moreover, the correction in that way is visualized as referred to both the lines and not only to l. 12.
Col. II 21-22. The asteriskos is between the two lines.
Col. II 23-24. The lines are in ekthesis, but if I try to use the indicated codification, I have problem in the visualization of the text.
 (Roberta Carlesimo)
 - Vote - Accept-Straight-to-Finalization - Review text carefully; dots have all been mis-entered. Metadata will also require a close examination before finalization (C. Michael Sampson)
 - Finalized - This is good to go (C. Michael Sampson)