Text given according to the edition (in l. 8 I have substituted the supplement [Tag] with [.?])
Also, just now while editing this text I have become aware of the possibility to signal indented and outdented lines in the text. I have tried to follow the instructions to indent l. 7, but I have failed and gave up (or maybe was it just not shown in the preview but would have appear correctly in the final version?). 

 - Created from SoSOL template

Signed-off-by: James Cowey <james.cowey@urz.uni-heidelberg.de>