Update revisionDesc
 - Submit - Text given according to the edition (in l. 8 I have substituted the supplement [Tag] with [.?])

Also, just now while editing this text I have become aware of the possibility to signal indented and outdented lines in the text. I have tried to follow the instructions to indent l. 7, but I have failed and gave up (or maybe was it just not shown in the preview but would have appear correctly in the final version?).  (Riccardo Vecchiato)
 - Vote - Accept-Straight-to-Finalization - lin8 [Ἁ]θὺ[ρ .?] => [Ἁ]θὺ[ρ] [.?] (James Cowey)
 - Finalized - Ready. (James Cowey)