new text In line one of the recto, the edition has a stauros. I changed it to a rho-cross which is clearly visible on the papyrus. The editor seems to use two different symbols for the stauros, so perhaps they were confunded and originally meant to differentiate stauros and rho-cross.
also there seem to be two stauroi at the end of the recto which have been omitted in the edition. 

on the verso in line 16, I added a diaeresis on the iota of υιου which has been omitted in the edition, but seems to be visible on the papyrus.
There are some other traces on the verso also missing in the edition.

 - Created from SoSOL template

Signed-off-by: Lajos Berkes <berkes.lajos@googlemail.com>