Update revisionDesc
 - Submit - new text
In line one of the recto, the edition has a stauros. I changed it to a rho-cross which is clearly visible on the papyrus. The editor seems to use two different symbols for the stauros, so perhaps they were confunded and originally meant to differentiate stauros and rho-cross.

also there seem to be two stauroi at the end of the recto which have been omitted in the edition. 

on the verso in line 16, I added a diaeresis on the iota of υιου which has been omitted in the edition, but seems to be visible on the papyrus.
There are some other traces on the verso also missing in the edition. (Conrad Theilemann)
 - Vote - Accept-Straight-to-Finalization - lin1: encode the change to the rho-cross as PN corr? (Cf. the submission note.)
lin2: regularize  ⲫⲱⲃ.
lin3: <#ⲙⲛⲧⲥ̄ⲛⲟⲟⲩⲥ=12#>
lin5: (γί(νεται)) => (γί(νονται)) [pace the edition, keep this as a tacit change); =><#ιβ=12#>.
lin6: => (μ(ηνὶ))
lin7: =>  Ἐνὼχ 
Verso: renumber the lines beginning with 1.
lin11: ϣⲡ.2ⲟ.1 =>ϣⲣ.2ⲟ.1 (Lajos Berkes)
 - Finalized - ready (Lajos Berkes)