Update revisionDesc
 - Submit - I have entered the right date both in the text (l. 17, following BL 7 239) and in the Duke Databank heading. 
I have also corrected the demotic of Antinoopolis l. 5. The correct reading is Μεγαλεισίου, as already implied but not formally stated in Susan Fogarty and Paul Schubert, Abstract from the Register of Epikrisis of the Ephebes of Antinoopolis, ZPE 224 (2022) p. 240. It is hard to make out the letters in the plate of Pap. Flor. XXX (or the photograph provided on the website of the PSI), which is why I have dotted the last three letters, but that is the only possible reading.  (François Gerardin)
 - Vote - ReturnToSender - Μεγαλε̣ι̣σ̣[ίου] (in place of Μεγαλιαν̣[) is not a possible reading. The letter after lambda is iota, and then alpha looks plausible. Next voter. (Nikolaos Gonis)
 - Vote - Return-Straight-to-Sender - Use the "corr" tag, I think. (James Cowey)
 - Submit - I am resubmitting this text to introduce into the digital edition at l. 5 a suggestion made by Vitelli in the first edition (n. 5 p. 109) but not recorded in the DDbDP.   (François Gerardin)
 - Vote - Accept-Straight-to-Finalization - Looks good. (James Cowey)
 - Finalized - Ready. (James Cowey)