Update revisionDesc
 - Submit - Entered from the edition. Please, indent line 31. 
In line 2 the editor put a superlinear dot and a dieresis on υιος, in line 11 a superlinear dot on υπερ. I did not find a way to represent the dot. 
In line 21, I did not remove the superlinear stroke from the old edition. In line 25, I removed the accent on Μεσορη, according to prof. Gonis edition. 
I left the superlinear stroke for the number, I don’t know if it’s correct. 
And I did nothing about the commentary (see: v spc, ecc). (Francesca Iacono)
 - Vote - Accept-Straight-to-Finalization - Should be checked carefully against the edition. Looks good. (James Cowey)
 - Finalized - Ready. (James Cowey)