Update revisionDesc
 - Commit - Provided an editorial correction in l. 14, μ[ .2]ενης to μ[ηχ]α̣νῆς (Myrto Malouta)
 - Submit - Corrected μ[  ̣  ̣]ενης to μ[ηχ]α̣νῆς in l. 14.
The context clearly suggests a water-lifting machine is what is being referred to (see e.g. P.Berl.Leihg. I 23.7-13). This is also supported by the pitch and nails mentioned in the text, which are used for machine construction and repair (see e.g. P.Lond. III 1177.226 for pitch and 241 for nails). 
If the reading α]πο is correct, then the text can be interpreted as “I shall water the arable land with the help of a water-lifting machine that will be built with materials (nails and pitch survive in the text) provided by the landlord, while I the tenant shall be responsible for part repairs and general maintenance”. That the materials will be provided by the landlord is not specifically stated in what survives, but can be understood by the juxtaposition to what the tenant states to be his responsibility (της δε) and this division of responsibility is a standard feature of land leases that include a water-lifting device, though the terms of who is responsible for what tend to vary.  (Myrto Malouta)
 - Vote - AcceptText - The suggestion is interesting. One might even try μ[η]χ̣α̣νῆς (a low trace might belong to chi, but the image doesn’t allow ambitious readings—ask for a proper one?), but the resulting sequence, ἀ]πὸ τῆς μ[ηχ]α̣νῆς ἐκ καινοῦ κατασκευῆς, is not smooth Greek. The problem is καινοῦ: read ἐκ καινῆς̣, which yields a phrase with some close parallels. We would have a new irrigator, to be built with (materials). (To say ‘I shall water the arable land with the help of a water-lifting machine’ doesn’t do justice to what precedes μ[η]χ̣α̣νῆς, but this is not an interpretative note.) (Nikolaos Gonis)
 - Vote - AcceptText - Suggestion looks good. I would print μ[ηχ]α̣νῆς, following the initial submission, because the traces below seem to be part of the series of dots b/t ll. 14 and 15 noted by the editor (the loose frag can be slightly rotated and contains the final nu of ἥλων in l. 15). As for the reading καινῆς̣ in l. 14, I see rather καινη corr. ex καινο, and I’m not sure there’s space for the sigma, even if the papyrus is a tad scrunched up here. In any case, I agree that ἐκ καινῆς was intended, as in l. 11 (though there the phrase appears to stand alone, without a noun). The text could use more attention in general: for starters, line ends are preserved, which allows all supplements to be pushed to the left and helps with the estimation of line lengths. For now, a minor improvement to l. 5: it’s a chi before υριαινης, which leads to Ἰ]σ̣χυριαίνης, a common name around Ptolemais Hormou in the Petaus archive (also change reading of ed. pr. from Ἀπιαίνησ to ]απιαίνης). (Graham Claytor)
 - Finalized - Ready. I have incorporated both suggestions made by Claytor. In lin5 I decided to leave the name Ἀπιαίνης (as entered in the DDbDP), since it is listed as such in the index to the volume. I also accepted the suggestion καινη corr. ex καινο, although I am not entirely convinced. I have adjusted some of the wrongly placed brackets. μη[νὸς -ca.?- ] in lin6 should probably be μη|[νὸς -ca.?- ] since there is plenty of free space at the end of the line on the papyrus where letters could have been written. (James Cowey)