Update chr.mitt.6.xml
For cross-references problems, this papyrus displays wrongly in the Navigator. M.Chr. 6 is a reprint of P.Petr. II 18 (1) and of P.Petr. III 32 c; neither of them has a DDB entry; I created the cross-references as file-filler just to avoid M.Chr. to aggregate with other TMs and display wrongly.